Add macros: route one control source to multiple parameters per track
Every synth, MIDI, audio and kit (affect-entire) track gets 4 macros. A
macro takes one source - a learned MIDI CC, a gold knob, or its own
automation lane - and drives up to 8 destinations, each scaled onto its
own from/to range (from > to inverts, from == to sends a constant).
Destinations are output CCs on MIDI clips, sound params on synths, and
global FX params on audio clips and kits. A macro can also feed a
higher-numbered macro to build compound controls.

Only one macro drives a given destination at a time: the earliest active
macro owns it and any other assignment is shadowed, flagged with a
blinking mod-button LED. Macro lanes show up in automation view like any
other param; editing a lane bakes it into the destinations' lanes, and
BACK undoes the lane and its destinations together.

Shift + vertical encoder in clip view toggles macro mode: the left four
mod buttons select macros, the gold knobs perform the selected one, and
holding a macro button turns the grid into a target picker - tap pads or
dial the select encoder to assign destinations (including params with no
pad), shape from/to with the gold knobs, shift+save to remove. Shift +
macro button jumps to the lane editor, where the 8 mod buttons quick-edit
that macro's targets.

Gated behind a community feature toggle. Serialization is additive: a
song that never touched macros writes byte-identical XML, and songs with
macros load fine on older firmware (unknown tags are skipped, macro lanes
dropped).

+++ src/deluge/modulation/midi/midi_param.h
@@ -23,6 +23,8 @@ class MIDIParam {
 public:
 	MIDIParam();
 
-	uint8_t cc;
+	// int32_t (not uint8_t) so the MIDIParamVector's 32-bit key can hold macro-lane pseudo-CCs (128-131)
+	// without the 8-bit-signed key sign-extending them into negatives. Real CCs are still 0..123.
+	int32_t cc;
 	AutoParam param;
 };
